omsMAXv.cpp: Casting between unsigned long and pointer
On most system the size of a pointer is the same as the size of "long".
One exception is Win64, where a pointer uses 64 bits and the "long"
is 32 bits.
This code is (as today) not used on a Windows 64 bit platform,
so silence the compiler warnings for this platforms by casting a pointer
to size_t and then "down" to "unsigned long".
This allows continous integration to pass.